+++ src/components/Recipe/RecipeItem/RecipeItem.tsx
@@ -1,5 +1,4 @@
-import { BottomSheet, Skeleton, useBottomSheet } from '@fun-eat/design-system';
-import type { MouseEventHandler } from 'react';
+import { Skeleton } from '@fun-eat/design-system';
 import { memo, useState } from 'react';
 import { Link } from 'react-router-dom';
 
@@ -10,14 +9,13 @@ import {
   recipeAuthor,
   recipeContent,
   recipeImage,
-  recipeProductWrapper,
   recipeTitle,
 } from './recipeItem.css';
 import RecipeFavoriteButton from '../RecipeFavoriteButton/RecipeFavoriteButton';
 import RecipeProductButton from '../RecipeProductButton/RecipeProductButton';
 
-import { ProductOverviewList } from '@/components/Product';
 import { RECIPE_CARD_DEFAULT_IMAGE_URL } from '@/constants/image';
+import { PATH } from '@/constants/path';
 import type { MemberRecipe, Recipe } from '@/types/recipe';
 
 interface RecipeItemProps {
@@ -26,21 +24,14 @@ interface RecipeItemProps {
 }
 
 const RecipeItem = ({ recipe, isMemberPage = false }: RecipeItemProps) => {
-  const { id, image, title, content, favorite, products } = recipe;
-  const { isOpen, isClosing, handleOpenBottomSheet, handleCloseBottomSheet } = useBottomSheet();
-
-  const author = 'author' in recipe ? recipe.author : null;
-
   const [isImageLoading, setIsImageLoading] = useState(true);
 
-  const handleOpenProductSheet: MouseEventHandler<HTMLDivElement> = (event) => {
-    event.preventDefault();
-    handleOpenBottomSheet();
-  };
+  const { id, image, title, content, favorite, products } = recipe;
+  const author = 'author' in recipe ? recipe.author : null;
 
   return (
     <>
-      <Link to={`${id}`}>
+      <Link to={`${PATH.RECIPE}/${id}`}>
         {!isMemberPage && (
           <div className={imageWrapper}>
             <img
@@ -54,8 +45,8 @@ const RecipeItem = ({ recipe, isMemberPage = false }: RecipeItemProps) => {
             <div className={favoriteButtonWrapper} onClick={(e) => e.preventDefault()}>
               <RecipeFavoriteButton recipeId={id} favorite={favorite} />
             </div>
-            <div className={productButtonWrapper} onClick={(e) => handleOpenProductSheet(e)}>
-              <RecipeProductButton isTranslucent />
+            <div className={productButtonWrapper} onClick={(e) => e.preventDefault()}>
+              <RecipeProductButton isTranslucent products={products} />
             </div>
           </div>
         )}
@@ -64,12 +55,6 @@ const RecipeItem = ({ recipe, isMemberPage = false }: RecipeItemProps) => {
         <p className={recipeAuthor}>{author && `${author.nickname} 님`}</p>
         <p className={recipeContent}>{content}</p>
       </Link>
-
-      <BottomSheet isOpen={isOpen} isClosing={isClosing} maxWidth="400px" close={handleCloseBottomSheet}>
-        <div className={recipeProductWrapper}>
-          <ProductOverviewList products={products} />
-        </div>
-      </BottomSheet>
     </>
   );
 };
